% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbeta.r
\name{rtbeta}
\alias{rtbeta}
\alias{dtbeta}
\alias{ptbeta}
\alias{qtbeta}
\title{Truncated beta distribution functions}
\usage{
rtbeta(n, alpha, beta, a = 0, b = 1)

dtbeta(x, alpha, beta, a = 0, b = 1)

ptbeta(q, alpha, beta, a = 0, b = 1)

qtbeta(p, alpha, beta, a = 0, b = 1)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{alpha}{vector of alpha parameters.}

\item{beta}{vector of beta parameters.}

\item{a}{vector of lower truncation limits}

\item{b}{vector of upper truncation limits}

\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\value{
dtbeta gives the density, ptbeta gives the distribution function, qtbeta gives the quantile function, and rtbeta generates random deviates.
}
\description{
pdf, cdf, inverse cdf, and random deviates of the truncated beta distribution.
}
\examples{
set.seed(1976)
rtbeta(10, 3, 4, 0.2, 0.8)
dtbeta(0.5, 3, 4, 0.2, 0.7)
qtbeta(0.1, 3, 4, 0.2, 0.9)
ptbeta(0.3, 3, 4, 0.2, 0.9)
rtbeta(5, 1, 2, a=0.5, b=0.9)
dtbeta(seq(0, 1, by=0.1), 1, 2, a=0.5, b=0.9)
ptbeta(seq(0, 1, by=0.1), 1, 2, a=0.5, b=0.9)
qtbeta(seq(0, 1, by=0.1), 1, 2, a=0.5, b=0.9)
}
