% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.r
\name{rtexp}
\alias{rtexp}
\alias{dtexp}
\alias{ptexp}
\alias{qtexp}
\title{Truncated exponential distribution functions}
\usage{
rtexp(n, rate, a = 0, b = Inf)

dtexp(x, rate, a = 0, b = Inf)

ptexp(q, rate, a = 0, b = Inf)

qtexp(p, rate, a = 0, b = Inf)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{rate}{vector of rate parameters.}

\item{a}{vector of lower truncation limits}

\item{b}{vector of upper truncation limits}

\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\value{
dtexp gives the density, ptexp gives the distribution function, qtexp gives the quantile function, and rtexp generates random deviates.
}
\description{
pdf, cdf, inverse cdf, and random deviates of the truncated Exponential distribution.
}
\examples{
rtexp(5, 1, a=0.5, b=3.0)
dtexp(seq(0, 4, by=0.5), 1, a=0.5, b=3.0)
ptexp(seq(0, 4, by=0.5), 1, a=0.5, b=3.0)
qtexp(seq(0, 1, by=0.1), 1, a=0.5, b=3.0)
}
