% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgamma.r
\name{rtgamma}
\alias{rtgamma}
\alias{dtgamma}
\alias{ptgamma}
\alias{qtgamma}
\title{Truncated gamma distribution functions}
\usage{
rtgamma(n, shape, scale = 1, a = 0, b = Inf)

dtgamma(x, shape, scale = 1, a = 0, b = Inf)

ptgamma(q, shape, scale = 1, a = 0, b = Inf)

qtgamma(p, shape, scale = 1, a = 0, b = Inf)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{shape}{vector of shape parameters.}

\item{scale}{vector of scale parameters.}

\item{a}{vector of lower truncation limits}

\item{b}{vector of upper truncation limits}

\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\value{
dtgamma gives the density, ptgamma gives the distribution function, qtgamma gives the quantile function, and rtgamma generates random deviates.
}
\description{
pdf, cdf, inverse cdf, and random deviates of the truncated gamma distribution.
}
\examples{
rtgamma(5, 1, scale=2, a=0.5, b=3.0)
dtgamma(seq(0, 4, by=0.5), 1, scale=2, a=0.5, b=3.0)
ptgamma(seq(0, 4, by=0.5), 1, scale=2, a=0.5, b=3.0)
qtgamma(seq(0, 1, by=0.1), 1, scale=2, a=0.5, b=3.0)
}
