% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlnorm.r
\name{rtlnorm}
\alias{rtlnorm}
\alias{dtlnorm}
\alias{ptlnorm}
\alias{qtlnorm}
\title{Truncated lognormal distribution functions}
\usage{
rtlnorm(n, meanlog = 0, sdlog = 1, a = 0, b = Inf)

dtlnorm(x, meanlog = 0, sdlog = 1, a = 0, b = Inf)

ptlnorm(q, meanlog = 0, sdlog = 1, a = 0, b = Inf)

qtlnorm(p, meanlog = 0, sdlog = 1, a = 0, b = Inf)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{meanlog}{vector of means on the natural log scale.}

\item{sdlog}{vector of standard deviations on the natural log scale.}

\item{a}{vector of lower truncation limits}

\item{b}{vector of upper truncation limits}

\item{x, q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\value{
dtlnorm gives the density, ptlnorm gives the distribution function, qtlnorm gives the quantile function, and rtlnorm generates random deviates.
}
\description{
pdf, cdf, inverse cdf, and random deviates of the truncated lognormal distribution.
}
\examples{
 rtlnorm(5, 1, 2, 0.5, 3.0)
 dtlnorm(seq(0, 4, by=0.5), 1, 2, 0.5, 3.0)
 ptlnorm(seq(0, 4, by=0.5), 1, 2, 0.5, 3.0)
 qtlnorm(seq(0, 1, by=0.1), 1, 2, 0.5, 3.0)
}
