% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncatedMeanVar.r
\name{meanTruncated}
\alias{meanTruncated}
\alias{varTruncated}
\title{Mean and Variance of a truncated distribution}
\usage{
meanTruncated(f, param1, param2, a, b)

varTruncated(f, param1, param2, a, b)
}
\arguments{
\item{f}{string name of a distribution (gamma, norm, beta, lnorm, weibull)}

\item{param1}{the first distribution parameter}

\item{param2}{the second distribution parameter}

\item{a}{lower truncation limit}

\item{b}{upper truncation limit}
}
\value{
a list containing the mean or variance and absolute error
}
\description{
Mean and Variance of a truncated distribution
}
\examples{
meanTruncated("gamma", 1, 2, 0.1, 5)
varTruncated("gamma", 1, 2, 0.1, 5)
}
