\name{TruncatedDistributions-package}
\alias{TruncatedDistributions-package}
\alias{TruncatedDistributions}
\docType{package}
\title{
  Truncated Distributions
}
\description{
  Contains distribution functions for many truncated distributions
}
\details{
  \tabular{ll}{
    Package: \tab TruncatedDistributions\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2013/03/15\cr
    License: \tab GPL (>=v2)\cr
  }
}
\author{
Rob Carnell <bertcarnell@gmail.com>
}
\keyword{distribution}
